fix: resolve CMake compile warnings (#43)
* fix: resolve CMake compile warnings

* style: apply clang-format with updated formatting rules

- Update .clang-format with stricter rules for templates, parameters, and lambdas
- Add clang-format pre-commit hook
- Reformat all C++ source files

* refactor: update clang-tidy checks and remove obsolete CI scripts

+++ tests/index/nndescent_test.cpp
@@ -43,8 +43,8 @@ class NnDescentTest : public ::testing::Test {
 
     // Init the vector data.
     srand(time(nullptr));
-    for (int i = 0; i < max_node_; ++i) {
-      for (int j = 0; j < dim_; ++j) {
+    for (uint32_t i = 0; i < max_node_; ++i) {
+      for (uint32_t j = 0; j < dim_; ++j) {
         data_[i * dim_ + j] = rand() % max_node_;
       }
     }
@@ -149,7 +149,7 @@ TEST_F(NnDescentSearchTest, SimpleSearchTest) {
   const size_t kSearchThreadNum = 16;
   std::vector<std::thread> tasks(kSearchThreadNum);
 
-  auto search_knn = [&](int i) {
+  auto search_knn = [&](uint32_t i) {
     for (; i < query_num_; i += kSearchThreadNum) {
       std::vector<uint32_t> ids(topk);
       auto cur_query = queries_.data() + i * dim_;
@@ -165,11 +165,11 @@ TEST_F(NnDescentSearchTest, SimpleSearchTest) {
     }
   };
 
-  for (int i = 0; i < kSearchThreadNum; i++) {
+  for (size_t i = 0; i < kSearchThreadNum; i++) {
     tasks[i] = std::thread(search_knn, i);
   }
 
-  for (int i = 0; i < kSearchThreadNum; i++) {
+  for (size_t i = 0; i < kSearchThreadNum; i++) {
     if (tasks[i].joinable()) {
       tasks[i].join();
     }
@@ -179,9 +179,9 @@ TEST_F(NnDescentSearchTest, SimpleSearchTest) {
 
   // Computing recall;
   size_t cnt = 0;
-  for (int i = 0; i < query_num_; i++) {
-    for (int j = 0; j < topk; j++) {
-      for (int k = 0; k < topk; k++) {
+  for (uint32_t i = 0; i < query_num_; i++) {
+    for (size_t j = 0; j < topk; j++) {
+      for (size_t k = 0; k < topk; k++) {
         if (res_pool[i][j] == answers_[i * gt_col_ + k]) {
           cnt++;
           break;
